feat: Turned the onboarding pages into a component. Added redirection after sign up
I turned the onboarding pages into a component so that we can still see the underlying pages. I added redirection after a user signs up so that they land on the page that they were previously trying to access.
